Records and PBs

+++ app/src/main/java/com/android/jared/linden/timingtrials/data/PersonalBest.kt
@@ -2,4 +2,4 @@ package com.android.jared.linden.timingtrials.data
 
 import org.threeten.bp.OffsetDateTime
 
-data class PersonalBest(val courseId: Long, val timeTrialId: Long?, val courseName: String, val distance: Long? = null, val millisTime: Long, val dateTime: OffsetDateTime? = null)
+data class PersonalBest(val courseId: Long, val timeTrialId: Long?, val courseName: String, val distance: Double? = null, val millisTime: Long, val dateTime: OffsetDateTime? = null)
